+++ HSCFiscalRegistrar/HSCFiscalRegistrar/DTO/Fiscalization/OFD/Sum.cs
@@ -2,7 +2,7 @@ namespace HSCFiscalRegistrar.DTO.Fiscalization.OFD
 {
     public class Sum
     {
-        public int Bills { get; set; }
-        public int Coins { get; set; }
+        public decimal Bills { get; set; }
+        public decimal Coins { get; set; }
     }
 }
